fix pivot
