+++ frontend/src/metabase/documents/hooks/use-card-data.ts
@@ -9,7 +9,7 @@ import { useSelector } from "metabase/redux";
 import type { UseCardDataResult } from "metabase/rich_text_editing/tiptap/EditorHost";
 import { getMetadata } from "metabase/selectors/metadata";
 import Question from "metabase-lib/v1/Question";
-import { getPivotOptions } from "metabase-lib/v1/queries/utils/pivot";
+import { getPivotOptions } from "metabase-lib/v1/queries/utils/pivot-options";
 import type { Card, Dataset, RawSeries } from "metabase-types/api";
 import { isObject } from "metabase-types/guards";
